Move LocationSelectWidget to locations
